+++ bindings/matrix-sdk-ffi/CHANGELOG.md
@@ -103,6 +103,8 @@ All notable changes to this project will be documented in this file.
 - [**breaking**] The `SpaceService`'s `joined_spaces` method has been renamed
   `top_level_joined_spaces` and `subscribe_to_joined_spaces` to `space_service.subscribe_to_top_level_joined_spaces`
   ([#5972](https://github.com/matrix-org/matrix-rust-sdk/pull/5972))
+- [**breaking**] `HumanQrGrantLoginError::UnableToCreateDevice` has been removed
+  ([#6141](https://github.com/matrix-org/matrix-rust-sdk/pull/6141)
 
 ## [0.16.0] - 2025-12-04